src/requirements.txt
adjustText==0.7.3
anykeystore==0.2
apex==0.9.10.dev0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.0.6
cryptacular==1.6.2
cycler==0.11.0
defusedxml==0.7.1
einops==0.5.0
filelock==3.8.2
fonttools==4.38.0
greenlet==2.0.0.post0
h5py==3.7.0
huggingface-hub==0.10.1
hupper==1.10.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.0.2
mkl-service==2.3.0
mpmath==1.2.1
networkx==3.0rc1
numpy==1.24.0rc2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-dali-cuda110==1.19.0
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
PasteDeploy==3.0.1
pbkdf2==1.3
Pillow==9.3.0
plaster==1.0
plaster-pastedeploy==0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.0.9
pyramid==2.0
pyramid-mailer==0.15.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytorch-ranger==0.1.1
pytorch-ssim==0.1
pytorch-warmup==0.1.1
pytz==2022.6
PyYAML==6.0
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
seaborn==0.12.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
SQLAlchemy==1.4.43
sympy==1.11.1
timm==0.6.11
torch==1.13.1
torch-optimizer==0.3.0
torchaudio==0.13.0
torchinfo==1.7.1
torchtriton==2.0.0+0d7e753227
torchvision==0.14.1+cu116
tqdm==4.64.1
transaction==3.0.1
translationstring==1.4
typing_extensions==4.4.0
urllib3==1.26.12
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.7
WTForms==3.0.1
wtforms-recaptcha==0.3.2
zope.deprecation==4.4.0
zope.interface==5.5.1
zope.sqlalchemy==1.6
nni
